% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSTCM.R
\name{SRSTCM}
\alias{SRSTCM}
\title{SRSTCM main_func.}
\usage{
SRSTCM(
  data,
  cancer,
  type,
  survival,
  multi.sample = "FALSE",
  minGene = 500,
  maxGene = 5000,
  maxUMI = 40000,
  pctMT = 20,
  id.type = "S",
  cell.line = "no",
  ngene.chr = 5,
  win.size = 25,
  KS.cut = 0.15,
  distance = "euclidean",
  n.cores = 1
)
}
\arguments{
\item{data}{raw data matrix; if the data is scRNA-seq, genes in rows; cell names in columns; if the data is bulk, genes in rows; sample names in columns.}

\item{cancer}{short for cancer: "LUNG";"LIHC";"BRCA";"OVCA";"CRC";SKCM".}

\item{type}{if the data are scRNA-seq, put "sc"; if the data are bulk, put "bulk".}

\item{survival}{survival data; if the survival data is not empty, survival data is a data frame; sample names, OS, and OS.time in columns.}

\item{multi.sample}{if scRNA-seq is a single sample, put "FALSE"; if scRNA-seq is multi-sample integration, put "TRUE".}

\item{minGene}{the minimum number of genes in a cell.}

\item{maxGene}{the maximum number of genes in a cell.}

\item{maxUMI}{the maximum UMI count in a cell.}

\item{pctMT}{mitochondrial gene ratio.}

\item{id.type}{gene id type: Symbol or Ensemble.}

\item{cell.line}{if the data are from pure cell line,put "yes"; if cell line data are a mixture of tumor and normal cells, still put "no".}

\item{ngene.chr}{minimal number of genes per chromosome for cell filtering.}

\item{win.size}{minimal window sizes for segmentation.}

\item{KS.cut}{segmentation parameters, input 0 to 1; larger looser criteria.}

\item{distance}{distance methods include euclidean, and correlation converted distance include pearson and spearman.}

\item{n.cores}{number of cores for parallel computing.}
}
\value{
if data is scRNA-seq: 1) scRNA-seq quality test results; 2) scRNA-seq cluster map; 3) aneuploid/diploid prediction results; 4)senescent tumor cells prediction results; 5)annotated cluster diagram of senescent tumor cells; 6)differential genes in senescent and non-senescent tumor cell clusters; 7)expression of the top 10 differentially expressed genes in senescent and non-senescent tumor cell clusters; if data is bulk: 1)the senescence score results; 2)survival outcomes in the high and low senescence score groups.
}
\description{
SRSTCM main_func.
}
\examples{
test.SRSTCM <- SRSTCM(data = data, cancer = "LUNG", type = "sc", survival = survival, multi.sample = "FALSE", minGene = 500, maxGene = 5000, maxUMI = 40000, pctMT = 20, id.type = "S", cell.line = "no", ngene.chr = 5,  win.size = 25, KS.cut = 0.15, distance = "euclidean", n.cores = 1)
}
